requirements.txt
pandas
scikit-learn
numpy
ipykernel
Flask
PyYAML

-e .
mlflow==2.2.2 